rule out the UB problem, and move the functions impl out of the .hpp file
